+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD (Unreleased)
 * Upgrade to v2.9.0 of the Docker Terraform Provider
+* Improve build and push output [#251](https://github.com/pulumi/pulumi-docker/pull/251)
 
 ---